# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quranify API - Quran memorization tracking platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  admin: "admin",
  mentor: "mentor",
  member: "member",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UpdateProfileBody {
  name?: string;
  email?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  mentor: "mentor",
  member: "member",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  halaqahGroup?: number | null;
  halaqahName?: string | null;
  totalAyat: number;
  streak: number;
  badges: string[];
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export type HafalanStatus = (typeof HafalanStatus)[keyof typeof HafalanStatus];

export const HafalanStatus = {
  New: "New",
  Review: "Review",
  Strong: "Strong",
} as const;

export interface Hafalan {
  id: number;
  userId: number;
  surah: string;
  surahNumber: number;
  ayahStart: number;
  ayahEnd: number;
  ayahCount: number;
  status: HafalanStatus;
  date: string;
  createdAt: string;
}

export type CreateHafalanBodyStatus =
  (typeof CreateHafalanBodyStatus)[keyof typeof CreateHafalanBodyStatus];

export const CreateHafalanBodyStatus = {
  New: "New",
  Review: "Review",
  Strong: "Strong",
} as const;

export interface CreateHafalanBody {
  surah: string;
  surahNumber: number;
  ayahStart: number;
  ayahEnd: number;
  status: CreateHafalanBodyStatus;
  date: string;
}

export type UpdateHafalanBodyStatus =
  (typeof UpdateHafalanBodyStatus)[keyof typeof UpdateHafalanBodyStatus];

export const UpdateHafalanBodyStatus = {
  New: "New",
  Review: "Review",
  Strong: "Strong",
} as const;

export interface UpdateHafalanBody {
  surah?: string;
  surahNumber?: number;
  ayahStart?: number;
  ayahEnd?: number;
  status?: UpdateHafalanBodyStatus;
  date?: string;
}

export interface Halaqah {
  id: number;
  name: string;
  mentorId: number;
  mentorName: string;
  memberCount: number;
  totalAyat: number;
  createdAt: string;
}

export interface HalaqahDetail {
  id: number;
  name: string;
  mentorId: number;
  mentorName: string;
  members: User[];
  totalAyat: number;
  createdAt: string;
}

export interface CreateHalaqahBody {
  name: string;
  mentorId: number;
}

export interface AssignHalaqahBody {
  halaqahId: number | null;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  totalAyat: number;
  streak: number;
  halaqahName?: string | null;
  badges: string[];
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  currentUserRank?: number | null;
}

export interface DashboardSummary {
  totalAyat: number;
  streak: number;
  lastActivity?: string | null;
  todayAyah: string;
  todayAyahReference: string;
  motivationalQuote: string;
  motivationalQuoteReference: string;
  recentHafalan: Hafalan[];
  badges: string[];
  weeklyGoal: number;
  weeklyProgress: number;
}

export interface WeeklyData {
  day: string;
  ayat: number;
}

export interface MonthlyData {
  week: string;
  ayat: number;
}

export interface AnalyticsResponse {
  weekly: WeeklyData[];
  monthly: MonthlyData[];
  totalThisWeek: number;
  totalThisMonth: number;
  averagePerDay: number;
}

export type ListHafalanParams = {
  surah?: string;
  status?: ListHafalanStatus;
  dateFrom?: string;
  dateTo?: string;
};

export type ListHafalanStatus =
  (typeof ListHafalanStatus)[keyof typeof ListHafalanStatus];

export const ListHafalanStatus = {
  New: "New",
  Review: "Review",
  Strong: "Strong",
} as const;