# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quranify API - Quran memorization tracking platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: hafalan
    description: Hafalan (memorization) tracking
  - name: leaderboard
    description: Leaderboard and rankings
  - name: halaqah
    description: Halaqah group management
  - name: users
    description: User management
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /hafalan:
    get:
      operationId: listHafalan
      tags: [hafalan]
      summary: List hafalan entries for current user
      parameters:
        - name: surah
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [New, Review, Strong]
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of hafalan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hafalan"
    post:
      operationId: createHafalan
      tags: [hafalan]
      summary: Create a new hafalan entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHafalanBody"
      responses:
        "201":
          description: Hafalan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hafalan"

  /hafalan/{id}:
    put:
      operationId: updateHafalan
      tags: [hafalan]
      summary: Update a hafalan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHafalanBody"
      responses:
        "200":
          description: Hafalan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hafalan"
    delete:
      operationId: deleteHafalan
      tags: [hafalan]
      summary: Delete a hafalan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get global leaderboard
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"

  /halaqah:
    get:
      operationId: listHalaqah
      tags: [halaqah]
      summary: List all halaqah groups
      responses:
        "200":
          description: List of halaqah groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Halaqah"
    post:
      operationId: createHalaqah
      tags: [halaqah]
      summary: Create a halaqah group (admin/mentor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHalaqahBody"
      responses:
        "201":
          description: Halaqah created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Halaqah"

  /halaqah/{id}:
    get:
      operationId: getHalaqah
      tags: [halaqah]
      summary: Get a halaqah group with members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Halaqah group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HalaqahDetail"

  /halaqah/{id}/leaderboard:
    get:
      operationId: getHalaqahLeaderboard
      tags: [halaqah]
      summary: Get leaderboard for a specific halaqah
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Halaqah leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary for current user
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/analytics:
    get:
      operationId: getAnalytics
      tags: [dashboard]
      summary: Get weekly and monthly analytics for current user
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{id}/assign-halaqah:
    put:
      operationId: assignHalaqah
      tags: [users]
      summary: Assign user to halaqah (admin/mentor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignHalaqahBody"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, mentor, member]
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, mentor, member]
        halaqahGroup:
          type: integer
          nullable: true
        halaqahName:
          type: string
          nullable: true
        totalAyat:
          type: integer
        streak:
          type: integer
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, totalAyat, streak, badges, createdAt]

    Hafalan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        surah:
          type: string
        surahNumber:
          type: integer
        ayahStart:
          type: integer
        ayahEnd:
          type: integer
        ayahCount:
          type: integer
        status:
          type: string
          enum: [New, Review, Strong]
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, userId, surah, surahNumber, ayahStart, ayahEnd, ayahCount, status, date, createdAt]

    CreateHafalanBody:
      type: object
      properties:
        surah:
          type: string
        surahNumber:
          type: integer
        ayahStart:
          type: integer
        ayahEnd:
          type: integer
        status:
          type: string
          enum: [New, Review, Strong]
        date:
          type: string
          format: date
      required: [surah, surahNumber, ayahStart, ayahEnd, status, date]

    UpdateHafalanBody:
      type: object
      properties:
        surah:
          type: string
        surahNumber:
          type: integer
        ayahStart:
          type: integer
        ayahEnd:
          type: integer
        status:
          type: string
          enum: [New, Review, Strong]
        date:
          type: string
          format: date

    Halaqah:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mentorId:
          type: integer
        mentorName:
          type: string
        memberCount:
          type: integer
        totalAyat:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, mentorId, mentorName, memberCount, totalAyat, createdAt]

    HalaqahDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mentorId:
          type: integer
        mentorName:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        totalAyat:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, mentorId, mentorName, members, totalAyat, createdAt]

    CreateHalaqahBody:
      type: object
      properties:
        name:
          type: string
        mentorId:
          type: integer
      required: [name, mentorId]

    AssignHalaqahBody:
      type: object
      properties:
        halaqahId:
          type: integer
          nullable: true
      required: [halaqahId]

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        totalAyat:
          type: integer
        streak:
          type: integer
        halaqahName:
          type: string
          nullable: true
        badges:
          type: array
          items:
            type: string
      required: [rank, userId, name, totalAyat, streak, badges]

    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        currentUserRank:
          type: integer
          nullable: true
      required: [entries]

    DashboardSummary:
      type: object
      properties:
        totalAyat:
          type: integer
        streak:
          type: integer
        lastActivity:
          type: string
          nullable: true
        todayAyah:
          type: string
        todayAyahReference:
          type: string
        motivationalQuote:
          type: string
        motivationalQuoteReference:
          type: string
        recentHafalan:
          type: array
          items:
            $ref: "#/components/schemas/Hafalan"
        badges:
          type: array
          items:
            type: string
        weeklyGoal:
          type: integer
        weeklyProgress:
          type: integer
      required: [totalAyat, streak, todayAyah, todayAyahReference, motivationalQuote, motivationalQuoteReference, recentHafalan, badges, weeklyGoal, weeklyProgress]

    WeeklyData:
      type: object
      properties:
        day:
          type: string
        ayat:
          type: integer
      required: [day, ayat]

    MonthlyData:
      type: object
      properties:
        week:
          type: string
        ayat:
          type: integer
      required: [week, ayat]

    AnalyticsResponse:
      type: object
      properties:
        weekly:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyData"
        monthly:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyData"
        totalThisWeek:
          type: integer
        totalThisMonth:
          type: integer
        averagePerDay:
          type: number
      required: [weekly, monthly, totalThisWeek, totalThisMonth, averagePerDay]