# lib/api-zod/src/generated/types/createHalaqahBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quranify API - Quran memorization tracking platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateHalaqahBody {
  name: string;
  mentorId: number;
}