# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quranify API - Quran memorization tracking platform
 * OpenAPI spec version: 0.1.0
 */
import type { Hafalan } from "./hafalan";

export interface DashboardSummary {
  totalAyat: number;
  streak: number;
  lastActivity?: string | null;
  todayAyah: string;
  todayAyahReference: string;
  motivationalQuote: string;
  motivationalQuoteReference: string;
  recentHafalan: Hafalan[];
  badges: string[];
  weeklyGoal: number;
  weeklyProgress: number;
}